requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==4.4.2
distro==1.9.0
exceptiongroup==1.2.0
ffmpeg==1.4
ffmpeg-python==0.2.0
Flask==3.0.2
future==1.0.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
moviepy==1.0.3
numpy
openai==1.14.0
Pillow==9.5.0
proglog==0.1.10
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
uuid